# apps/API/DTOs/Orders/CreateMutualFundOrderDto.cs
namespace API.DTOs.Orders;

public class CreateMutualFundOrderDto
{

    public int AccountId { get; set; }
    public int AssetId { get; set; }
    public long Quantity { get; set; }
    public string OperationTypeId { get; set; } = null!;
}